gabi merged without conflict, now to combine

+++ app/src/main/java/com/example/myapplication/Produce.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 
 public class Produce implements Serializable {
     private String name;
-    private double price;
+    private Float price;
 
-    public Produce(String name, double price) {
+    public Produce(String name, Float price) {
         this.name = name;
         this.price = price;
     }
@@ -14,6 +14,6 @@ public class Produce implements Serializable {
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
 
-    public double getPrice() { return price; }
-    public void setPrice(double price) { this.price = price; }
+    public Float getPrice() { return price; }
+    public void setPrice(Float price) { this.price = price; }
 }
